# Port Join operator to WorkProcessor model 