--- a/presto-main/src/main/java/io/prestosql/operator/FilterAndProjectOperator.java
+++ b/presto-main/src/main/java/io/prestosql/operator/FilterAndProjectOperator.java
@@ -21,6 +21,7 @@ import io.prestosql.memory.context.LocalMemoryContext;
 import io.prestosql.memory.context.MemoryTrackingContext;
 import io.prestosql.operator.WorkProcessorOperatorAdapter.AdapterWorkProcessorOperator;
 import io.prestosql.operator.WorkProcessorOperatorAdapter.AdapterWorkProcessorOperatorFactory;
+import io.prestosql.operator.WorkProcessorOperatorAdapter.ProcessorContext;
 import io.prestosql.operator.project.PageProcessor;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.type.Type;
@@ -142,12 +143,12 @@ public class FilterAndProjectOperator
         }
 
         @Override
-        public AdapterWorkProcessorOperator create(Session session, MemoryTrackingContext memoryTrackingContext, DriverYieldSignal yieldSignal)
+        public AdapterWorkProcessorOperator create(ProcessorContext processorContext)
         {
             return new FilterAndProjectOperator(
-                    session,
-                    memoryTrackingContext,
-                    yieldSignal,
+                    processorContext.getSession(),
+                    processorContext.getMemoryTrackingContext(),
+                    processorContext.getDriverYieldSignal(),
                     Optional.empty(),
                     processor.get(),
                     types,
@@ -174,12 +175,12 @@ public class FilterAndProjectOperator
         }
 
         @Override
-        public WorkProcessorOperator create(Session session, MemoryTrackingContext memoryTrackingContext, DriverYieldSignal yieldSignal, WorkProcessor<Page> sourcePages)
+        public WorkProcessorOperator create(ProcessorContext processorContext, WorkProcessor<Page> sourcePages)
         {
             return new FilterAndProjectOperator(
-                    session,
-                    memoryTrackingContext,
-                    yieldSignal,
+                    processorContext.getSession(),
+                    processorContext.getMemoryTrackingContext(),
+                    processorContext.getDriverYieldSignal(),
                     Optional.of(sourcePages),
                     processor.get(),
                     types,