--- a/presto-main/src/main/java/io/prestosql/operator/TopNOperator.java
+++ b/presto-main/src/main/java/io/prestosql/operator/TopNOperator.java
@@ -14,11 +14,11 @@
 package io.prestosql.operator;
 
 import com.google.common.collect.ImmutableList;
-import io.prestosql.Session;
 import io.prestosql.memory.context.MemoryTrackingContext;
 import io.prestosql.operator.WorkProcessor.TransformationState;
 import io.prestosql.operator.WorkProcessorOperatorAdapter.AdapterWorkProcessorOperator;
 import io.prestosql.operator.WorkProcessorOperatorAdapter.AdapterWorkProcessorOperatorFactory;
+import io.prestosql.operator.WorkProcessorOperatorAdapter.ProcessorContext;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.block.SortOrder;
 import io.prestosql.spi.type.Type;
@@ -103,13 +103,11 @@ public class TopNOperator
 
         @Override
         public WorkProcessorOperator create(
-                Session session,
-                MemoryTrackingContext memoryTrackingContext,
-                DriverYieldSignal yieldSignal,
+                ProcessorContext processorContext,
                 WorkProcessor<Page> sourcePages)
         {
             return new TopNOperator(
-                    memoryTrackingContext,
+                    processorContext.getMemoryTrackingContext(),
                     Optional.of(sourcePages),
                     sourceTypes,
                     n,
@@ -118,13 +116,10 @@ public class TopNOperator
         }
 
         @Override
-        public AdapterWorkProcessorOperator create(
-                Session session,
-                MemoryTrackingContext memoryTrackingContext,
-                DriverYieldSignal yieldSignal)
+        public AdapterWorkProcessorOperator create(ProcessorContext processorContext)
         {
             return new TopNOperator(
-                    memoryTrackingContext,
+                    processorContext.getMemoryTrackingContext(),
                     Optional.empty(),
                     sourceTypes,
                     n,