--- a/presto-main/src/test/java/io/prestosql/operator/TestWorkProcessorPipelineSourceOperator.java
+++ b/presto-main/src/test/java/io/prestosql/operator/TestWorkProcessorPipelineSourceOperator.java
@@ -24,6 +24,7 @@ import io.prestosql.metadata.Split;
 import io.prestosql.operator.WorkProcessor.Transformation;
 import io.prestosql.operator.WorkProcessor.TransformationState;
 import io.prestosql.operator.WorkProcessorAssertion.Transform;
+import io.prestosql.operator.WorkProcessorOperatorAdapter.ProcessorContext;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.connector.UpdatablePageSource;
 import io.prestosql.sql.planner.plan.PlanNodeId;
@@ -407,7 +408,7 @@ public class TestWorkProcessorPipelineSourceOperator
         }
 
         @Override
-        public WorkProcessorOperator create(Session session, MemoryTrackingContext memoryTrackingContext, DriverYieldSignal yieldSignal, WorkProcessor<Page> sourcePages)
+        public WorkProcessorOperator create(ProcessorContext processorContext, WorkProcessor<Page> sourcePages)
         {
             assertNull(operator, "source operator already created");
             operator = new TestWorkProcessorOperator(sourcePages.transform(transformation));