TO-DO:

- Track memory for the output pages

The benchmark that we took 

```
                                                                                                                                                                 Normal                         Work Processor
Benchmark                                              (buildHashEnabled)  (buildRowsRepetition)  (hashColumns)  (matchRate)  (outputColumns)  Mode  Cnt     Score     Error  Units          Score     Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1        varchar          N/A              N/A  avgt   30   853.215 ±   5.996  ms/op       843.781 ±   7.716  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1         bigint          N/A              N/A  avgt   30   531.294 ±  49.266  ms/op       480.013 ±  11.462  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1            all          N/A              N/A  avgt   30  1314.879 ± 266.532  ms/op       955.585 ±  21.856  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5        varchar          N/A              N/A  avgt   30  1967.050 ± 113.097  ms/op      1170.089 ±  24.498  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5         bigint          N/A              N/A  avgt   30  1251.594 ± 190.671  ms/op       584.872 ±   4.955  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5            all          N/A              N/A  avgt   30  2322.659 ± 153.450  ms/op      1385.845 ±  15.290  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1        varchar          N/A              N/A  avgt   30  1137.780 ±  77.309  ms/op      1168.215 ± 110.874  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1         bigint          N/A              N/A  avgt   30  1098.345 ± 141.290  ms/op      1066.497 ± 112.083  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1            all          N/A              N/A  avgt   30  1437.359 ±  86.187  ms/op      1400.772 ±  86.195  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5        varchar          N/A              N/A  avgt   30  1801.792 ± 133.888  ms/op      1768.608 ± 122.690  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5         bigint          N/A              N/A  avgt   30  1274.933 ± 219.359  ms/op      1508.048 ± 232.207  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5            all          N/A              N/A  avgt   30  2133.180 ± 173.386  ms/op      2083.435 ± 156.605  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar          0.1           bigint  avgt   30   501.720 ±  40.658  ms/op       491.646 ±  38.148  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar          0.1              all  avgt   30   508.512 ±  37.853  ms/op       515.810 ±  38.215  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            1           bigint  avgt   30   495.618 ±  31.399  ms/op       491.188 ±  36.070  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            1              all  avgt   30   608.514 ±  34.263  ms/op       548.757 ±  36.449  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            2           bigint  avgt   30   307.822 ±  22.860  ms/op       310.354 ±  24.534  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            2              all  avgt   30   413.530 ±  27.973  ms/op       368.702 ±  29.651  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint          0.1           bigint  avgt   30   385.541 ±  31.138  ms/op       427.413 ±  42.924  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint          0.1              all  avgt   30   418.381 ±  33.469  ms/op       456.264 ±  46.562  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            1           bigint  avgt   30   393.607 ±  28.436  ms/op       376.974 ±  29.006  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            1              all  avgt   30   534.679 ±  32.480  ms/op       471.156 ±  35.624  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            2           bigint  avgt   30   238.572 ±  16.595  ms/op       222.459 ±  15.790  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            2              all  avgt   30   351.910 ±  25.507  ms/op       314.410 ±  23.851  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all          0.1           bigint  avgt   30   530.216 ±  45.057  ms/op       547.886 ±  42.634  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all          0.1              all  avgt   30   549.515 ±  42.296  ms/op       566.152 ±  47.411  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            1           bigint  avgt   30   566.055 ±  44.590  ms/op       521.172 ±  32.902  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            1              all  avgt   30   639.596 ±  38.572  ms/op       590.736 ±  41.050  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            2           bigint  avgt   30   347.534 ±  28.444  ms/op       333.202 ±  24.433  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            2              all  avgt   30   465.093 ±  33.190  ms/op       412.810 ±  29.516  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar          0.1           bigint  avgt   30   453.407 ±  35.255  ms/op       462.998 ±  41.880  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar          0.1              all  avgt   30   590.261 ±  48.365  ms/op       593.472 ±  46.046  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            1           bigint  avgt   30   862.731 ±  64.805  ms/op       828.970 ±  39.277  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            1              all  avgt   30  1382.688 ±  81.018  ms/op      1335.846 ±  69.668  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            2           bigint  avgt   30   610.187 ±  54.426  ms/op       606.609 ±  42.457  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            2              all  avgt   30  1212.495 ±  74.364  ms/op      1149.922 ±  69.049  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint          0.1           bigint  avgt   30   332.809 ±  31.834  ms/op       428.480 ±  35.487  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint          0.1              all  avgt   30   463.897 ±  37.050  ms/op       571.712 ±  42.395  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            1           bigint  avgt   30   754.607 ±  42.370  ms/op       751.556 ±  48.769  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            1              all  avgt   30  1312.524 ±  81.300  ms/op      1249.487 ±  74.263  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            2           bigint  avgt   30   538.278 ±  31.206  ms/op       350.649 ±   2.183  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            2              all  avgt   30  1125.858 ±  70.048  ms/op       735.775 ±  10.836  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all          0.1           bigint  avgt   30   498.699 ±  36.658  ms/op       320.271 ±   9.060  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all          0.1              all  avgt   30   614.061 ±  43.807  ms/op       400.066 ±   5.442  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            1           bigint  avgt   30   948.549 ±  67.885  ms/op       576.193 ±  19.730  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            1              all  avgt   30  1429.589 ±  74.592  ms/op       951.784 ±   5.846  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            2           bigint  avgt   30   675.079 ±  49.582  ms/op       489.229 ±  60.943  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            2              all  avgt   30  1236.805 ±  65.670  ms/op       782.930 ±   5.811  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar          0.1           bigint  avgt   30   353.206 ±  28.255  ms/op       208.393 ±   6.011  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar          0.1              all  avgt   30   385.564 ±  27.942  ms/op       314.950 ±  62.543  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            1           bigint  avgt   30   430.317 ±  29.014  ms/op       425.125 ±  38.157  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            1              all  avgt   30   559.318 ±  33.943  ms/op       500.482 ±  30.443  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            2           bigint  avgt   30   278.476 ±  19.800  ms/op       268.309 ±  21.216  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            2              all  avgt   30   379.907 ±  27.119  ms/op       336.251 ±  23.976  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint          0.1           bigint  avgt   30   338.865 ±  34.612  ms/op       341.070 ±  29.223  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint          0.1              all  avgt   30   378.213 ±  29.146  ms/op       375.125 ±  32.809  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            1           bigint  avgt   30   383.541 ±  26.826  ms/op       366.838 ±  30.932  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            1              all  avgt   30   536.607 ±  36.405  ms/op       478.893 ±  32.060  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            2           bigint  avgt   30   239.812 ±  16.928  ms/op       227.144 ±  16.707  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            2              all  avgt   30   362.953 ±  28.234  ms/op       312.567 ±  24.701  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all          0.1           bigint  avgt   30   456.443 ±  38.262  ms/op       478.295 ±  35.756  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all          0.1              all  avgt   30   451.874 ±  39.621  ms/op       459.029 ±  46.079  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            1           bigint  avgt   30   496.606 ±  37.871  ms/op       407.739 ±  62.407  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            1              all  avgt   30   617.664 ±  39.957  ms/op       344.157 ±   3.700  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            2           bigint  avgt   30   309.336 ±  20.785  ms/op       190.537 ±   0.883  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            2              all  avgt   30   420.864 ±  29.838  ms/op       330.008 ±  47.625  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar          0.1           bigint  avgt   30   324.342 ±  24.455  ms/op       333.935 ±  28.268  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar          0.1              all  avgt   30   453.015 ±  41.154  ms/op       474.684 ±  41.022  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            1           bigint  avgt   30   798.051 ±  45.413  ms/op       785.330 ±  45.606  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            1              all  avgt   30  1305.680 ±  75.774  ms/op      1276.600 ±  65.511  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            2           bigint  avgt   30   592.777 ±  44.939  ms/op       574.000 ±  42.446  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            2              all  avgt   30  1222.249 ±  75.990  ms/op      1140.042 ±  63.089  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint          0.1           bigint  avgt   30   316.400 ±  25.455  ms/op       357.719 ±  52.560  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint          0.1              all  avgt   30   454.575 ±  33.161  ms/op       502.640 ±  48.392  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            1           bigint  avgt   30   755.437 ±  43.953  ms/op       719.082 ±  42.112  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            1              all  avgt   30  1332.002 ±  81.092  ms/op      1270.086 ±  77.205  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            2           bigint  avgt   30   531.291 ±  40.300  ms/op       518.361 ±  46.572  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            2              all  avgt   30  1146.824 ±  62.395  ms/op      1093.630 ±  65.947  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all          0.1           bigint  avgt   30   463.740 ±  37.975  ms/op       461.982 ±  34.699  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all          0.1              all  avgt   30   536.401 ±  52.768  ms/op       578.423 ±  43.188  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            1           bigint  avgt   30   866.933 ±  50.954  ms/op       857.101 ±  50.179  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            1              all  avgt   30  1417.945 ± 100.627  ms/op      1327.941 ±  75.018  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            2           bigint  avgt   30   667.043 ±  38.734  ms/op       614.397 ±  38.889  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            2              all  avgt   30  1209.488 ±  66.812  ms/op      1154.332 ±  69.110  ms/op

```