@sopel39 2019-08-19
Could you add support for operator info since https://github.com/prestosql/presto/pull/1292 landed