@sopel39 2019-09-30
Thanks. I think it looks great. Few comments:
* We don't track input page memory usage when join operator is blocked waiting for build side to create. I don't think it's a big deal though so we could skip it for now.
* Could you add some `BenchmarkLookupJoinOperator` (similar as `BenchmarkScanFilterAndProjectOperator`). This is pretty important operator, but we don't have JMH benchmarks for it.