@sopel39 2019-10-16
My cross check benchmarks. I see no perf regression
```
BEFORE
Benchmark                                              (buildHashEnabled)  (buildRowsRepetition)  (hashColumns)  (matchRate)  (outputColumns)  Mode  Cnt     Score    Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1        varchar          N/A              N/A  avgt   30   772.912 ± 17.507  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1         bigint          N/A              N/A  avgt   30   599.097 ±  7.977  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1            all          N/A              N/A  avgt   30   919.520 ± 16.951  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5        varchar          N/A              N/A  avgt   30  1189.149 ± 15.586  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5         bigint          N/A              N/A  avgt   30   963.000 ± 19.728  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5            all          N/A              N/A  avgt   30  1469.037 ± 29.790  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1        varchar          N/A              N/A  avgt   30   589.513 ± 10.992  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1         bigint          N/A              N/A  avgt   30   593.725 ± 16.187  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1            all          N/A              N/A  avgt   30   756.048 ±  9.447  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5        varchar          N/A              N/A  avgt   30  1098.474 ± 25.568  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5         bigint          N/A              N/A  avgt   30   937.327 ± 25.069  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5            all          N/A              N/A  avgt   30  1275.358 ± 27.665  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar          0.1           bigint  avgt   30   279.608 ±  7.899  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar          0.1              all  avgt   30   290.244 ±  7.535  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            1           bigint  avgt   30   299.395 ±  3.219  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            1              all  avgt   30   391.116 ±  6.416  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            2           bigint  avgt   30   232.341 ±  5.677  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            2              all  avgt   30   320.221 ±  3.875  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint          0.1           bigint  avgt   30   244.150 ±  5.529  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint          0.1              all  avgt   30   263.375 ± 12.851  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            1           bigint  avgt   30   231.694 ±  6.450  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            1              all  avgt   30   340.814 ±  4.839  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            2           bigint  avgt   30   169.877 ±  3.451  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            2              all  avgt   30   291.666 ±  2.015  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all          0.1           bigint  avgt   30   304.997 ±  5.323  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all          0.1              all  avgt   30   325.450 ± 11.985  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            1           bigint  avgt   30   343.881 ±  4.730  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            1              all  avgt   30   439.131 ±  5.431  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            2           bigint  avgt   30   270.582 ±  7.432  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1            all            2              all  avgt   30   359.956 ±  3.180  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar          0.1           bigint  avgt   30   297.857 ±  3.700  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar          0.1              all  avgt   30   412.051 ±  3.122  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            1           bigint  avgt   30   621.612 ± 10.255  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            1              all  avgt   30  1131.548 ± 18.941  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            2           bigint  avgt   30   529.287 ± 16.839  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5        varchar            2              all  avgt   30  1054.403 ± 16.351  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint          0.1           bigint  avgt   30   215.835 ±  0.988  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint          0.1              all  avgt   30   349.575 ±  4.510  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            1           bigint  avgt   30   541.646 ± 12.641  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            1              all  avgt   30  1094.250 ± 13.889  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            2           bigint  avgt   30   473.873 ±  2.722  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5         bigint            2              all  avgt   30   998.368 ± 13.439  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all          0.1           bigint  avgt   30   321.826 ±  3.308  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all          0.1              all  avgt   30   431.876 ±  7.525  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            1           bigint  avgt   30   664.797 ± 19.505  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            1              all  avgt   30  1155.691 ± 18.261  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            2           bigint  avgt   30   571.474 ±  7.998  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      5            all            2              all  avgt   30  1060.077 ± 16.819  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar          0.1           bigint  avgt   30   201.476 ±  3.129  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar          0.1              all  avgt   30   208.981 ±  2.923  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            1           bigint  avgt   30   269.131 ±  5.101  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            1              all  avgt   30   362.766 ±  5.312  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            2           bigint  avgt   30   205.341 ±  3.508  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1        varchar            2              all  avgt   30   299.217 ±  2.624  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint          0.1           bigint  avgt   30   184.861 ±  3.936  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint          0.1              all  avgt   30   215.597 ±  4.662  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            1           bigint  avgt   30   236.406 ±  3.485  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            1              all  avgt   30   345.485 ±  7.625  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            2           bigint  avgt   30   175.598 ±  2.996  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1         bigint            2              all  avgt   30   298.822 ±  1.499  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all          0.1           bigint  avgt   30   244.783 ± 18.865  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all          0.1              all  avgt   30   264.736 ±  7.854  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            1           bigint  avgt   30   331.802 ±  4.767  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            1              all  avgt   30   407.144 ±  6.583  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            2           bigint  avgt   30   241.139 ±  3.490  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      1            all            2              all  avgt   30   336.196 ±  2.988  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar          0.1           bigint  avgt   30   222.213 ±  3.077  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar          0.1              all  avgt   30   339.723 ±  5.223  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            1           bigint  avgt   30   594.820 ± 14.358  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            1              all  avgt   30  1104.040 ± 13.630  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            2           bigint  avgt   30   507.414 ± 11.642  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5        varchar            2              all  avgt   30   993.278 ±  9.684  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint          0.1           bigint  avgt   30   215.390 ±  3.191  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint          0.1              all  avgt   30   335.985 ±  1.667  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            1           bigint  avgt   30   538.142 ± 11.598  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            1              all  avgt   30  1097.312 ± 21.946  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            2           bigint  avgt   30   446.214 ±  3.840  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5         bigint            2              all  avgt   30   997.632 ± 13.583  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all          0.1           bigint  avgt   30   282.484 ± 17.225  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all          0.1              all  avgt   30   402.091 ±  2.432  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            1           bigint  avgt   30   633.086 ± 20.578  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            1              all  avgt   30  1130.972 ± 14.219  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            2           bigint  avgt   30   559.543 ±  2.982  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                 true                      5            all            2              all  avgt   30  1038.818 ± 14.581  ms/op

AFTER
Benchmark                                              (buildHashEnabled)  (buildRowsRepetition)  (hashColumns)  (matchRate)  (outputColumns)  Mode  Cnt     Score    Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1        varchar          N/A              N/A  avgt   30   779.959 ± 17.032  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1         bigint          N/A              N/A  avgt   30   585.120 ± 14.536  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      1            all          N/A              N/A  avgt   30   902.129 ± 21.895  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5        varchar          N/A              N/A  avgt   30  1166.988 ± 44.285  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5         bigint          N/A              N/A  avgt   30  1003.722 ± 17.016  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false                      5            all          N/A              N/A  avgt   30  1428.198 ± 41.004  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1        varchar          N/A              N/A  avgt   30   584.140 ± 10.459  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1         bigint          N/A              N/A  avgt   30   581.254 ± 11.376  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      1            all          N/A              N/A  avgt   30   765.561 ±  9.916  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5        varchar          N/A              N/A  avgt   30  1043.553 ± 13.414  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5         bigint          N/A              N/A  avgt   30   938.467 ± 20.738  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true                      5            all          N/A              N/A  avgt   30  1259.402 ± 43.434  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar          0.1           bigint  avgt   30   275.365 ±  6.876  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar          0.1              all  avgt   30   286.175 ±  7.282  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            1           bigint  avgt   30   274.388 ±  6.982  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            1              all  avgt   30   354.059 ±  7.472  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            2           bigint  avgt   30   216.005 ±  1.432  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1        varchar            2              all  avgt   30   281.464 ±  2.695  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint          0.1           bigint  avgt   30   189.976 ±  6.486  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint          0.1              all  avgt   30   272.066 ±  7.277  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            1           bigint  avgt   30   228.318 ±  2.286  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            1              all  avgt   30   307.474 ± 10.010  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            2           bigint  avgt   30   162.113 ±  3.681  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash                false                      1         bigint            2              all  avgt   30   252.427 ±  2.431  ms/op
…